feat: 예약내역리스트 api 구현
예약내역리스트 api를 실제 동작하도록 구현함
- Repository 계층
  - 매개변수 자료형을 Long으로 변경함 (customerId를 Long으로 받기 때문)
  - 사용자 id로 예약 내역 호출하는 메서드 추가
- Controller 계층에서 현재 로그인 사용자의 식별자를 받아서 동작하도록 수정함
- Dto의 구성을 변경함 (리스트<객체> -> 객체)
- Service 계층
  - 더미 데이터 제거
  - DB로부터 읽어온 엔티티를 Dto로 변환 후 반환

+++ server/src/main/java/com/genesisairport/reservation/respository/CustomerRepository.java
@@ -11,5 +11,5 @@ public interface CustomerRepository extends JpaRepository<Customer, Long> {
 
     Optional<Customer> findByEmail(String email);
 
-    Customer findCustomerById(int id);
+    Customer findCustomerById(Long id);
 }
